had to use a cast due to incompatible types in for loop.

+++ source/tools.cpp
@@ -162,12 +162,12 @@ namespace libtp::tools
     int32_t getStageIndex( const char* stage )
     {
         // Find the index of this stage
-        for ( int32_t stageIDX = 0; stageIDX < sizeof( data::stage::allStages ) / sizeof( data::stage::allStages[0] );
+        for ( uint32_t stageIDX = 0; stageIDX < sizeof( data::stage::allStages ) / sizeof( data::stage::allStages[0] );
               stageIDX++ )
         {
             if ( strcmp( stage, libtp::data::stage::allStages[stageIDX] ) == 0 )
             {
-                return stageIDX;
+                return static_cast<int32_t>( stageIDX );
             }
         }
